+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/Response.kt
@@ -15,22 +15,11 @@ data class Response<T : Any>(
     val ok: Boolean = false,
     @Optional
     val description: String? = null,
-    @SerialName("migrate_to_chat_id")
-    @Optional
-    val migrateToChatId: Identifier? = null,
-    @SerialName("retry_after")
-    @Optional
-    val retryAfter: Int? = null,
     @SerialName("error_code")
     @Optional
     val errorCode: Int? = null,
     @Optional
-    val result: T? = null
-) {
-    @Transient
-    val waitUntil: DateTime? by lazy {
-        retryAfter ?.let {
-            DateTime.now().plus(TimeUnit.SECONDS.toMillis(it.toLong()))
-        }
-    }
-}
+    val result: T? = null,
+    @Optional
+    val parameters: ResponseParametersRaw? = null
+)